Introduce GetIBusKeyCodeString
We should send KeyCode oriented string to identify actual key code for degenerated case: e.g. "backslash" KeySym is from two different KeyCode "IntlRo" and "IntlYen" in Japanese Keyboard.

The identification should be meet DOM Level 4 Event spec. 
https://dvcs.w3.org/hg/d4e/raw-file/tip/source_respec.htm

BUG=169993
TEST=try bots.

Review URL: https://chromiumcodereview.appspot.com/12082055

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@180978 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: bac1eba1e9bc16809c56f09a16aba11311423527

+++ chrome/browser/chromeos/input_method/ibus_keymap.h
@@ -6,15 +6,16 @@
 #define CHROME_BROWSER_CHROMEOS_INPUT_METHOD_IBUS_KEYMAP_H_
 
 #include <string>
+#include "base/basictypes.h"
 
 namespace chromeos {
 namespace input_method {
 
-// Translate the key value from an IBus constant to a string.
+// Translates the key value from an IBus constant to a string.
 std::string GetIBusKey(int keyval);
 
-// Translate the unmodified keycode from an IBus constant to a string.
-std::string GetIBusKeyCode(int keycode);
+// Translates the unmodified keycode from an IBus constant to a string.
+std::string GetIBusKeyCode(uint16 keycode);
 
 }  // namespace input_method
 }  // namespace chromeos
